+++ examples/functions/compact_map/function.tf
@@ -1,0 +1,7 @@
+# Return map with null and empty values removed.
+provider::stdlib::compact_map({
+  "hello" = "world",
+  "foo"   = "",
+  "bar"   = null
+})
+# result => { "hello" = "world" }