+++ stdlib/map/compact_map_test.go
@@ -11,35 +11,21 @@ import (
 	mapfunc "github.com/mschuchard/terraform-provider-stdlib/stdlib/map"
 )
 
-func TestFlattenMapFunction(test *testing.T) {
+func TestCompactMapFunction(test *testing.T) {
 	test.Parallel()
 
 	standardTestCases := map[string]struct {
 		request  function.RunRequest
 		expected function.RunResponse
 	}{
-		"prepend": {
+		"standard": {
 			request: function.RunRequest{
 				Arguments: function.NewArgumentsData([]attr.Value{
-					types.ListValueMust(
-						types.MapType{ElemType: types.StringType},
-						[]attr.Value{types.MapValueMust(types.StringType, map[string]attr.Value{"hello": types.StringValue("world")}), types.MapValueMust(types.StringType, map[string]attr.Value{"foo": types.StringValue("bar")})},
-					),
+					types.MapValueMust(types.StringType, map[string]attr.Value{"hello": types.StringValue("world"), "foo": types.StringValue(""), "bar": types.StringNull()}),
 				}),
 			},
 			expected: function.RunResponse{
-				Result: function.NewResultData(types.MapValueMust(types.StringType, map[string]attr.Value{"hello": types.StringValue("world"), "foo": types.StringValue("bar")})),
-			},
-		},
-		"list-maps-length": {
-			request: function.RunRequest{
-				Arguments: function.NewArgumentsData([]attr.Value{
-					types.ListValueMust(types.MapType{ElemType: types.StringType}, []attr.Value{}),
-				}),
-			},
-			expected: function.RunResponse{
-				Error:  function.NewArgumentFuncError(0, "flatten_map: list of maps parameter must be at least length 1"),
-				Result: function.NewResultData(types.MapUnknown(types.StringType)),
+				Result: function.NewResultData(types.MapValueMust(types.StringType, map[string]attr.Value{"hello": types.StringValue("world")})),
 			},
 		},
 	}
@@ -50,7 +36,7 @@ func TestFlattenMapFunction(test *testing.T) {
 			result := function.RunResponse{Result: function.NewResultData(types.MapUnknown(types.StringType))}
 
 			// execute function and store result
-			mapfunc.NewFlattenMapFunction().Run(context.Background(), testCase.request, &result)
+			mapfunc.NewCompactMapFunction().Run(context.Background(), testCase.request, &result)
 
 			// compare results
 			if !result.Error.Equal(testCase.expected.Error) {