+++ examples/functions/coalesce_map/function.tf
@@ -1,0 +1,8 @@
+# Return first non-empty map.
+provider::stdlib::coalesce_map(
+  {},
+  { "hello" = "world" },
+  { "foo"   = "bar" },
+  {}
+}
+# result => { "hello" = "world" }