+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 2.1.0 (Next)
-- Add `factorial` function.
+- Add `compact_map` and `factorial` functions.
 
 ### 2.0.3
 - Implement `multiple` type custom functions.